Dashboard logic

+++ dashboard/src/main/kotlin/com/gigahack/dashboard/repository/DashboardRepository.kt
@@ -3,4 +3,4 @@ package com.gigahack.dashboard.repository
 import com.gigahack.dashboard.model.UserSettings
 import org.springframework.data.mongodb.repository.MongoRepository
 
-interface DashboardRepository : MongoRepository<UserSettings, Long>
+interface DashboardRepository : MongoRepository<UserSettings, Int>
